+++ src/lib/renderer/utils/colorspace/utils/round.ts
@@ -1,0 +1,6 @@
+
+
+export function roundWithPrecision(value: number, original: number): number {
+
+    return +value.toFixed(original.toString().split('.')[1]?.length ?? 0);
+}